core: frontend: refactor code for imu status
