+++ core/frontend/src/components/vehiclesetup/configuration/accelerometer/ArdupilotAccelerometerSetup.vue
@@ -66,6 +66,7 @@ import autopilot_data from '@/store/autopilot'
 import { Dictionary } from '@/types/common'
 import decode, { deviceId } from '@/utils/deviceid_decoder'
 
+import { imu_is_calibrated, imu_temperature_is_calibrated } from '../common'
 import QuickAccelerometerCalibration from './QuickAccelerometerCalibration.vue'
 
 export default Vue.extend({
@@ -93,42 +94,11 @@ export default Vue.extend({
         .filter((param) => param.value !== 0)
         .map((parameter) => decode(parameter.name, parameter.value))
     },
-    // TODO: refactor, same code as OnboardSensors.vue
     imu_is_calibrated(): Dictionary<boolean> {
-      const results = {} as Dictionary<boolean>
-      for (const imu of this.imus) {
-        const param_radix = imu.param.split('_ID')[0]
-        const offset_params_names = [`${param_radix}OFFS_X`, `${param_radix}OFFS_Y`, `${param_radix}OFFS_Z`]
-        const scale_params_names = [`${param_radix}SCAL_X`, `${param_radix}SCAL_Y`, `${param_radix}SCAL_Z`]
-        const offset_params = offset_params_names.map(
-          (name) => autopilot_data.parameter(name),
-        )
-        const scale_params = scale_params_names.map(
-          (name) => autopilot_data.parameter(name),
-        )
-        const is_at_default_offsets = offset_params.every((param) => param?.value === 0.0)
-        const is_at_default_scale = scale_params.every((param) => param?.value === 1.0)
-        results[imu.param] = offset_params.isEmpty() || scale_params.isEmpty()
-        || !is_at_default_offsets || !is_at_default_scale
-      }
-      return results
+      return imu_is_calibrated(this.imus, autopilot_data)
     },
     imu_temperature_is_calibrated(): Dictionary<{ calibrated: boolean, calibrationTemperature: number }> {
-      const results = {} as Dictionary<{ calibrated: boolean, calibrationTemperature: number }>
-      for (const imu of this.imus) {
-        let param_radix = imu.param.split('_ID')[0]
-        // CALTEMP parameters contains ID for the first sensor, _ID does not, so we need to add it
-        if (!/\d$/.test(param_radix)) {
-          param_radix += '1'
-        }
-        const name = `${param_radix}_CALTEMP`
-        const parameter = autopilot_data.parameter(name)
-        results[imu.param] = {
-          calibrated: parameter !== undefined && parameter.value !== -300,
-          calibrationTemperature: parameter?.value ?? 0,
-        }
-      }
-      return results
+      return imu_temperature_is_calibrated(this.imus, autopilot_data)
     },
   },
   methods: {
